requirements.txt
# pip install -r requirements.txt

# Base ----------------------------------------
cmake==3.26.3

matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scikit-build>=0.16.0
scipy==1.9.3
sqlalchemy>=2.0.0
vizer==0.1.5
urllib3==1.25.11
joblib==0.14.1
scikit-learn==1.0.2
tqdm>=4.41.0
torch==1.13.0
threadpoolctl==3.1.0
flask==2.0.3
flask-cors==3.0.1
pyzmq==25.0.0
pytest-benchmark==3.2.3
proglog==0.1.9
opencv-contrib-python>=4.5.0
face-recognition==1.3.0
face-recognition-models==0.3.0

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0